doubleVals 24.5

+++ DSL_Analyzer/ChartValues.cs
@@ -9,7 +9,7 @@ namespace DSL_Analyzer
     public class ChartValues
     {
         public string name { get; set; }
-        public List<int> Xvals { get; set; }
-        public List<int> Yvals { get; set; }
+        public List<double> Xvals { get; set; }
+        public List<double> Yvals { get; set; }
     }
 }
